[refactor] 테스트용 클래스 테스트 패키지 하위로 이동
