+++ src/test/java/study/step3/CarTest.java
@@ -4,7 +4,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import step3.common.FixNumberGenerator;
 import step3.model.Car;
 import step3.model.MovableStrategy;
